requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
kafka-python==2.0.1
oauthlib==3.1.0
py4j==0.10.7
PySocks==1.7.1
pyspark==2.4.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.8.0
urllib3==1.25.9
